Improve test logic moving test to provisioner
