+++ helper/tests/acc/provisioners.go
@@ -1,0 +1,126 @@
+package acc
+
+import (
+	"bytes"
+	"fmt"
+	"github.com/hashicorp/packer/helper/tests"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+
+	amazonEBS "github.com/hashicorp/packer/builder/amazon/ebs/acceptance"
+	virtualboxISO "github.com/hashicorp/packer/builder/virtualbox/iso/acceptance"
+	"github.com/hashicorp/packer/command"
+	"github.com/hashicorp/packer/packer"
+)
+
+func TestProvisionersAgainstBuilders(provisionerAcc ProvisionerAcceptance, t *testing.T) {
+	b := builderAccTestCheck(t)
+	provisioner := provisionerAcc.GetName()
+
+	// Get builders type to test provisioners against
+	var builders []string
+	for k := range BuildersAccTest {
+		// This will validate that only defined builders are executed against
+		if b != "all" && !strings.Contains(b, k) {
+			continue
+		}
+		builders = append(builders, k)
+	}
+
+	// build template file and run a build for each builder with the current provisioner
+	for _, builder := range builders {
+		builderAcc := BuildersAccTest[builder]
+		builderConfig, err := builderAcc.GetConfig()
+		if err != nil {
+			t.Fatalf("bad: failed to read builder config: %s", err.Error())
+		}
+
+		testName := fmt.Sprintf("testing %s builder against %s provisioner", builder, provisioner)
+		t.Run(testName, func(t *testing.T) {
+			provisionerConfig, err := provisionerAcc.GetConfig()
+			if err != nil {
+				t.Fatalf("bad: failed to read provisioner config: %s", err.Error())
+			}
+
+			// Write json template
+			out := bytes.NewBuffer(nil)
+			fmt.Fprintf(out, `{"builders": [%s],"provisioners": [%s]}`, builderConfig, provisionerConfig)
+			fileName := fmt.Sprintf("%s_%s.json", builder, provisioner)
+			filePath := filepath.Join("./", fileName)
+			writeJsonTemplate(out, filePath, t)
+
+			// set pre-config with necessary builder and provisioner
+			c := buildCommand(t, builderAcc, provisionerAcc)
+			args := []string{
+				filePath,
+			}
+
+			err = provisionerAcc.RunTest(c, args)
+			if err != nil {
+				// Cleanup created resources
+				testshelper.CleanupFiles(fileName)
+				builderAcc.CleanUp()
+				t.Fatalf("bad: failed to to run build: %s", err.Error())
+			}
+
+			// Cleanup created resources
+			testshelper.CleanupFiles(fileName)
+			err = builderAcc.CleanUp()
+			if err != nil {
+				t.Fatalf("bad: failed to clean up resources: %s", err.Error())
+			}
+		})
+	}
+}
+
+func builderAccTestCheck(t *testing.T) string {
+	b := os.Getenv("ACC_TEST_BUILDERS")
+	// validate if we want to run provisioners acc tests
+	if b == "" {
+		t.Skip("Provisioners Acceptance tests skipped unless env 'ACC_TEST_BUILDERS' is set")
+	}
+	return b
+}
+
+func writeJsonTemplate(out *bytes.Buffer, filePath string, t *testing.T) {
+	outputFile, err := os.Create(filePath)
+	if err != nil {
+		t.Fatalf("bad: failed to create template file: %s", err.Error())
+	}
+	_, err = outputFile.Write(out.Bytes())
+	if err != nil {
+		t.Fatalf("bad: failed to write template file: %s", err.Error())
+	}
+	outputFile.Sync()
+}
+
+func buildCommand(t *testing.T, builder BuilderAcceptance, provisioner ProvisionerAcceptance) *command.BuildCommand {
+	c := &command.BuildCommand{
+		Meta: testshelper.TestMetaFile(t),
+	}
+	c.CoreConfig.Components.BuilderStore = builder.GetBuilderStore()
+	c.CoreConfig.Components.ProvisionerStore = provisioner.GetProvisionerStore()
+
+	return c
+}
+
+type ProvisionerAcceptance interface {
+	GetName() string
+	GetConfig() (string, error)
+	GetProvisionerStore() packer.MapOfProvisioner
+	RunTest(c *command.BuildCommand, args []string) error
+}
+
+type BuilderAcceptance interface {
+	GetConfig() (string, error)
+	GetBuilderStore() packer.MapOfBuilder
+	CleanUp() error
+}
+
+// List of all builders available for acceptance test
+var BuildersAccTest = map[string]BuilderAcceptance{
+	"virtualbox-iso": new(virtualboxISO.VirtualBoxISOAccTest),
+	"amazon-ebs":     new(amazonEBS.AmazonEBSAccTest),
+}
