+++ provisioner/shell/test-fixtures/shell-provisioner.txt
